refactor(plan_SERVERSPLIT slice 7): the file-extraction bodies leave server.py
The nine `file-extraction` tool bodies move VERBATIM out of embedded/server.py
into embedded/tool_sections/file_extraction.py. Pure move: the only edits are
the dropped registration decorator (contract rule 2 — registration stays driven
from server.py's binding loop, once per execution of that module body, so the
canonical import, the bare-name spec load and the runpy __main__ load each keep
a full 94-tool app) and the rewrite of the singleton/helper reads to `rt.<name>`,
resolved at CALL time against the one patchable home (contract rule 3).

The nine were physically SPLIT in server.py — two above
`extract_complete_element_cdp` and five below it, because that
element-extraction body was misfiled among them. They close up here in their
registration (surface) order; the stray is relocated by slice 8.

The two goldens under this section's subsystem
(tests/goldens/file_based_structure_to_file.json,
extract_element_structure_list_convert.json) are driven by
tests/test_cloner_schemas.py against `file_based_element_cloner` DIRECTLY, one
layer below the tool bodies, so a tool-body move cannot reach them; the file is
re-run in this slice's lane and `git diff -- tests/goldens/` is empty.

server.py 2391 -> 2116 LOC (274 body lines plus the one-line
`file_based_element_cloner` alias import they were the last consumer of); the
GRANDFATHER cap ratchets DOWN to the measured post-ruff-format actual, and
tests/source_scan.py's floor ratchets UP 8 -> 9.

Verified: wire surface IDENTICAL to tests/goldens/tool_surface.json;
SECTION_TOOLS 94 with per-app [94, 94, 94] across the three identities;
`--list-sections` subprocess Total: 94 tools; unit lane 2330 passed / 1 skipped
(the -1 vs the 2332 base is section 5.5's parametrize shrinking by the pruned
alias, and the one failure is the load-sensitive
test_close_instance_offload.py::test_loop_stays_responsive_during_stuck_kill,
green on a re-run of its own file); real-Chrome test_e2e_data_tools.py 5 passed,
test_e2e_animations.py 8 passed, test_e2e_transport.py 1 passed; budgets, ruff
format+check, suppression owners, vulture, ty (81 warnings, 0 errors) and
gen_release_contract --check all clean.
