+++ tests/source_scan.py
@@ -31,9 +31,9 @@ from pathlib import Path
 #: which a dropped section module reads as a legitimately smaller set. Slice 1
 #: (``cookies_storage``) took it to 3, slice 2 (``tabs``) to 4, slice 3
 #: (``debugging``) to 5, slice 4 (``dynamic_hooks``) to 6, slice 5
-#: (``progressive_cloning``) to 7 and slice 6 (``network_debugging``) to 8;
-#: slice 11 ends at 13.
-MIN_TOOL_SOURCE_FILES = 8
+#: (``progressive_cloning``) to 7, slice 6 (``network_debugging``) to 8 and
+#: slice 7 (``file_extraction``) to 9; slice 11 ends at 13.
+MIN_TOOL_SOURCE_FILES = 9
 
 
 def collect_tool_source_files(server_mod, runtime_mod, section_modules, floor):
